etl_01/etl/requirements.txt
elasticsearch==7.7.0
psycopg2-binary==2.9.3
python-dotenv==0.20.0

